Separate userRegistry src into its own module
