+++ strato/core/strato-model/package.yaml
@@ -57,6 +57,7 @@ library:
     - path-pieces
     - persistent
     - primitive
+    - raw-strings-qq
     - relapse           # remove this
     - secp256k1-haskell
     - servant
